backend/requirements.txt
Flask
flask-cors
python-dotenv
redis
langchain
langchain-community
langchain-openai
langchain-huggingface
chromadb
sentence-transformers
